+++ src/main/scala/edu/ie3/simona/model/participant/load/FixedLoadModel.scala
@@ -8,6 +8,7 @@ package edu.ie3.simona.model.participant.load
 
 import edu.ie3.datamodel.models.input.system.LoadInput
 import edu.ie3.simona.config.RuntimeConfig.LoadRuntimeConfig
+import edu.ie3.simona.model.participant.ParticipantFlexModel
 import edu.ie3.simona.model.participant.ParticipantModel.{
   ActivePowerOperatingPoint,
   FixedState,
@@ -15,6 +16,11 @@ import edu.ie3.simona.model.participant.ParticipantModel.{
   ParticipantModelFactory,
 }
 import edu.ie3.simona.model.participant.control.QControl
+import edu.ie3.simona.model.participant.flex.{
+  ParticipantConstantEnergyLimitFlexModel,
+  ParticipantInflexiblePowerLimitFlexModel,
+}
+import edu.ie3.simona.ontology.messages.flex.FlexType
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.util.scala.quantities.ApparentPower
 import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
@@ -36,6 +42,12 @@ class FixedLoadModel(
     private val activePower: Power,
 ) extends LoadModel[FixedState]
     with ParticipantFixedState[ActivePowerOperatingPoint] {
+
+  override val flexModels: Map[FlexType, ParticipantFlexModel[FixedState]] =
+    Map(
+      FlexType.PowerLimit -> ParticipantInflexiblePowerLimitFlexModel(this),
+      FlexType.EnergyBoundaries -> ParticipantConstantEnergyLimitFlexModel(this),
+    )
 
   override def determineOperatingPoint(
       state: FixedState