+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented forecast for `LoadProfileService` [#1704](https://github.com/ie3-institute/simona/issues/1704)
 - Include necessary updates of version in docs and adaptions into `rtd/releaseProcess` [#1716](https://github.com/ie3-institute/simona/issues/1716)
 - Implemented energy boundaries flex options for EVCS [#1682](https://github.com/ie3-institute/simona/issues/1682)
+- Implemented energy boundaries flex options for load model [#1705](https://github.com/ie3-institute/simona/issues/1705)
+- Added secondary data provision handling in EmAgent [#1652](https://github.com/ie3-institute/simona/issues/1652)
 - Added `GridAgentCoordinator` [#1617](https://github.com/ie3-institute/simona/issues/1617)
 
 ### Changed